{"schema":"decodex/commit/1","summary":"Consolidate cohesive operator lane readout modules","authority":"manual"}
