--- apps/decodex-app/Sources/DecodexApp/OperatorLaneReadoutRow.swift
@@ -1,73 +1,0 @@
-import SwiftUI
-
-struct OperatorLaneReadoutRow: View {
-	let title: String
-	let items: [OperatorLaneReadoutItem]
-	let trailing: String?
-	@Environment(\.colorScheme) private var colorScheme
-
-	init(title: String, items: [OperatorLaneReadoutItem], trailing: String? = nil) {
-		self.title = title
-		self.items = items
-		self.trailing = trailing
-	}
-
-	var body: some View {
-		VStack(alignment: .leading, spacing: OperatorLaneReadoutLayout.itemRowSpacing) {
-			HStack(alignment: .firstTextBaseline, spacing: OperatorLaneReadoutLayout.columnSpacing) {
-				OperatorLaneReadoutLabelView(title: title)
-
-				if summaryFragments.isEmpty == false {
-					OperatorLaneReadoutSummaryView(fragments: summaryFragments)
-						.lineLimit(1)
-						.allowsTightening(true)
-						.fixedSize(horizontal: true, vertical: false)
-						.help(accessibilityText)
-				} else {
-					Spacer(minLength: 0)
-				}
-
-				if let trailing = panelTrimmed(trailing) {
-					Text(trailing)
-						.font(OperatorLanePopoverStyle.metaFont)
-						.foregroundStyle(OperatorLanePopoverStyle.mutedText(colorScheme))
-						.lineLimit(1)
-						.fixedSize(horizontal: true, vertical: false)
-				}
-			}
-		}
-		.fixedSize(horizontal: true, vertical: false)
-	}
-
-	private var summaryFragments: [[OperatorLaneReadoutTextRun]] {
-		let fragments = normalizedTitle == "tools"
-			? toolSummaryFragments
-			: items.map(\.summaryRuns)
-		return fragments.filter { $0.isEmpty == false }
-	}
-
-	private var normalizedTitle: String {
-		title.lowercased()
-	}
-
-	private var toolSummaryFragments: [[OperatorLaneReadoutTextRun]] {
-		var fragments = [[OperatorLaneReadoutTextRun]]()
-		if let calls = value(for: "tools") ?? value(for: "tool calls") {
-			fragments.append([.value(calls), .meta(" tools")])
-		}
-		if let maxOutput = value(for: "max output") ?? value(for: "max tool output") ?? value(for: "largest output") {
-			fragments.append([.value(maxOutput), .meta(" max output")])
-		}
-
-		return fragments
-	}
-
-	private func value(for label: String) -> String? {
-		items.first { $0.matchesLabel(label) }?.displayValue
-	}
-
-	private var accessibilityText: String {
-		items.map(\.accessibilityText)
-			.joined(separator: ", ")
-	}
-}
