--- apps/decodex-app/Sources/DecodexApp/OperatorLaneReadoutSummaryView.swift
@@ -1,58 +1,0 @@
-import SwiftUI
-
-struct OperatorLaneReadoutSummaryView: View {
-	let fragments: [[OperatorLaneReadoutTextRun]]
-	@Environment(\.colorScheme) private var colorScheme
-
-	var body: some View {
-		HStack(alignment: .firstTextBaseline, spacing: 0) {
-			ForEach(fragments.indices, id: \.self) { index in
-				OperatorLaneReadoutRunsView(runs: fragments[index])
-
-				if index != fragments.indices.last {
-					Text(" · ")
-						.font(OperatorLanePopoverStyle.separatorFont)
-						.foregroundStyle(OperatorLanePopoverStyle.mutedText(colorScheme))
-				}
-			}
-		}
-		.fixedSize(horizontal: true, vertical: false)
-	}
-}
-
-private struct OperatorLaneReadoutRunsView: View {
-	let runs: [OperatorLaneReadoutTextRun]
-	@Environment(\.colorScheme) private var colorScheme
-
-	var body: some View {
-		HStack(alignment: .firstTextBaseline, spacing: 0) {
-			ForEach(runs.indices, id: \.self) { index in
-				let run = runs[index]
-				Text(run.text)
-					.font(font(for: run.role))
-					.foregroundStyle(foreground(for: run.role))
-					.monospacedDigit()
-					.lineLimit(1)
-			}
-		}
-		.fixedSize(horizontal: true, vertical: false)
-	}
-
-	private func font(for role: OperatorLaneReadoutTextRole) -> Font {
-		switch role {
-		case .meta:
-			return OperatorLanePopoverStyle.metaFont
-		case .value:
-			return OperatorLanePopoverStyle.valueFont
-		}
-	}
-
-	private func foreground(for role: OperatorLaneReadoutTextRole) -> Color {
-		switch role {
-		case .meta:
-			return OperatorLanePopoverStyle.mutedText(colorScheme)
-		case .value:
-			return OperatorLanePopoverStyle.primaryText(colorScheme)
-		}
-	}
-}
