--- apps/decodex-app/Sources/DecodexApp/OperatorLaneReadoutTotals.swift
@@ -1,110 +1,0 @@
-import Foundation
-
-extension OperatorLanePopoverView {
-	var totalOverviewMetrics: [OperatorTotalMetric] {
-		guard let lifecycleMetrics = run.lifecycleMetrics else {
-			return []
-		}
-
-		return [
-			contextMetric(lifecycleMetrics),
-			trackerMetric(lifecycleMetrics),
-			protocolMetric(lifecycleMetrics),
-		].compactMap { $0 }
-	}
-
-	func contextMetric(_ metrics: OperatorLifecycleMetrics) -> OperatorTotalMetric? {
-		var items = [OperatorLaneReadoutItem]()
-		if metrics.inputTokensCumulative > 0 {
-			items.append(
-				OperatorLaneReadoutItem(
-					label: "input",
-					value: formatCompactCount(metrics.inputTokensCumulative)
-				)
-			)
-		}
-		if metrics.outputTokensCumulative > 0 {
-			items.append(
-				OperatorLaneReadoutItem(
-					label: "output",
-					value: formatCompactCount(metrics.outputTokensCumulative)
-				)
-			)
-		}
-
-		let current = metrics.inputTokensCurrent ?? activity?.inputTokensCurrent
-		let peak = metrics.inputTokensPeak ?? activity?.inputTokensMax
-		if let current {
-			items.append(
-				OperatorLaneReadoutItem(
-					label: "current window",
-					value: formatCompactCount(current)
-				)
-			)
-		}
-		if let peak, peak != current {
-			items.append(
-				OperatorLaneReadoutItem(label: "peak window", value: formatCompactCount(peak))
-			)
-		}
-
-		guard items.isEmpty == false else {
-			return nil
-		}
-
-		return OperatorTotalMetric(
-			title: "Context",
-			items: items
-		)
-	}
-
-	func trackerMetric(_ metrics: OperatorLifecycleMetrics) -> OperatorTotalMetric? {
-		guard let bucket = lifecycleBucket(named: "Tracker", in: metrics.buckets) else {
-			return nil
-		}
-
-		var items = [OperatorLaneReadoutItem]()
-		if bucket.eventCount > 0 {
-			items.append(OperatorLaneReadoutItem(label: "events", value: formatCompactCount(bucket.eventCount)))
-		}
-		if bucket.toolCallCount > 0 {
-			items.append(OperatorLaneReadoutItem(label: "tools", value: formatCompactCount(bucket.toolCallCount)))
-		}
-		if bucket.outputBytes > 0 {
-			items.append(OperatorLaneReadoutItem(label: "output bytes", value: formatCompactBytes(bucket.outputBytes)))
-		}
-
-		guard items.isEmpty == false else {
-			return nil
-		}
-
-		return OperatorTotalMetric(
-			title: "Tracker",
-			items: items
-		)
-	}
-
-	func protocolMetric(_ metrics: OperatorLifecycleMetrics) -> OperatorTotalMetric? {
-		guard metrics.protocolEventCount > 0 || metrics.childEventCount > 0 else {
-			return nil
-		}
-
-		var items = [OperatorLaneReadoutItem]()
-		if metrics.protocolEventCount > 0 {
-			items.append(OperatorLaneReadoutItem(label: "events", value: formatCompactCount(metrics.protocolEventCount)))
-		}
-		if metrics.childEventCount > 0 {
-			items.append(
-				OperatorLaneReadoutItem(
-					label: "child events",
-					value: formatCompactCount(metrics.childEventCount)
-				)
-			)
-		}
-
-		return OperatorTotalMetric(
-			title: "Protocol",
-			items: items
-		)
-	}
-}
